+++ app/services/DatasetService.scala
@@ -384,5 +384,6 @@ trait DatasetService {
 
   def incrementDownloads(id: UUID, user: Option[User])
 
-}
+  def getMetrics(user: Option[User]): Iterable[Dataset]
 
+}