+++ app/services/mongodb/MongoDBDatasetService.scala
@@ -1627,6 +1627,10 @@ class MongoDBDatasetService @Inject() (
       case None => {}
     }
   }
+
+  def getMetrics(user: Option[User]): Iterable[Dataset] = {
+    Dataset.find(MongoDBObject("stats" -> MongoDBObject("$exists" -> true), "trash" -> false)).toIterable
+  }
 }
 
 object Dataset extends ModelCompanion[Dataset, ObjectId] {